+++ src/gitstore/__init__.py
@@ -3,6 +3,7 @@ from .mirror import SyncDiff, RefChange
 from .fs import FS
 from .exceptions import StaleSnapshotError
 from .copy import copy_to_repo, copy_from_repo, copy_to_repo_dry_run, copy_from_repo_dry_run
+from .copy import remove_from_repo, remove_from_repo_dry_run
 from .copy import sync_to_repo, sync_from_repo, sync_to_repo_dry_run, sync_from_repo_dry_run
 from .copy import CopyReport, CopyPlan, CopyAction, CopyError, FileEntry, SyncPlan, SyncAction
 
@@ -10,6 +11,7 @@ __all__ = [
     "GitStore", "SyncDiff", "RefChange", "FS", "StaleSnapshotError",
     "copy_to_repo", "copy_from_repo",
     "copy_to_repo_dry_run", "copy_from_repo_dry_run",
+    "remove_from_repo", "remove_from_repo_dry_run",
     "sync_to_repo", "sync_from_repo",
     "sync_to_repo_dry_run", "sync_from_repo_dry_run",
     "CopyReport", "CopyPlan", "CopyAction", "CopyError", "FileEntry", "SyncPlan", "SyncAction",